fix z-score and print; rename script
